requirements.txt
tornado
kafka-python
pyspark
pymongo
